+++ docs/changelog.md
@@ -1,3 +1,9 @@
+## 1.4.0
+> Released Jan X, 2017
+
+- **NEW**: HeaderAnchor extension is now deprecated.  It will be removed in a future version.
+- **NEW**: HeaderAnchor is no longer included in the Github extension.
+
 ## 1.3.0
 > Released Jan 1, 2017